pyproject.toml
[project]
name = "sim-studentcup2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.6",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.5",
]
